requirements.txt
html2text
python-dotenv
CanvasScraper
canvasapi
pathvalidate